Allow 'notify' endpoint (separate from 'complete')
Some gateways expect different behaviour from the endpoint based on whether its sending a redirect for the user's browser, or a server-to-server failover gateway response. IP Payments for example expects a 2xx HTTP response on the server-to-server message, and fails on 3xx. We still need a 3xx redirect for the user URL though.

Always return 2xx codes on notify endpoint.
The server-to-server comms need to rely on HTTP codes to mark its requests as "accepted",
even if the actual payment failed. This is different from the user redirection URL.